+++ mg_custom_nodes/ZhiHui6_zhihui_nodes_comfyui_20260201/web/TextMergerNode-JS.js
@@ -64,6 +64,10 @@ app.registerExtension({
                     }
                 }
             };
+
+            if (inputcountWidget) {
+                 this.updateInputs(inputcountWidget.value);
+            }
         };
     }
 });