+++ mg_custom_nodes/ZhiHui6_zhihui_nodes_comfyui_20260201/CHANGELOG.md
@@ -1,4 +1,13 @@
 # 更新日志
+## 1.0.8 2026-2-1
+- 新增 提示词替换 节点(PromptReplace)：用于在提示词中替换指定文本。
+- 重构 文本合并 节点(TextMergerNode)，移除 user_text 输入，调整端口逻辑
+- 以及其他各项优化 
+    <br>
+- Added PromptReplace node: allows users to replace specified text in prompts.
+- Refactored TextMergerNode: removed user_text input and adjusted port logic
+- Various other optimizations
+
 ## 1.0.7 2026-1-30
 - 新增 字体设计师 节点(TypeDesigner)：用于设计和生成自定义字体样式。
 - 更新 图像宽高比节点(ImageAspectRatio)：为部分模型预设添加影院及标准分辨率